+++ src/main/java/domain/LotteryGame.java
@@ -1,11 +1,13 @@
 package domain;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import domain.generatestrategy.LotteryGenerateFamily;
+import domain.lottery.Lotteries;
+import domain.lottery.Lottery;
+import domain.lottery.WinningLottery;
 
 public class LotteryGame {
