+++ src/main/java/view/OutputView.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import domain.Lottery;
+import domain.lottery.Lottery;
 import domain.Rank;
 
 public class OutputView {
