+++ src/test/java/domain/lottery/LotteriesTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.lottery;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 
@@ -10,6 +10,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import domain.generatestrategy.LotteryRandomGeneratorStrategy;
+import domain.lottery.Lotteries;
 
 public class LotteriesTest {
