+++ src/test/java/domain/lottery/WinningLotteryTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.lottery;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 
@@ -13,6 +13,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import domain.Rank;
 
 public class WinningLotteryTest {
